requirements.txt
# Flask
flask==0.10.1

# Other
python-gcm==0.4

# Production
gunicorn==19.4.5